+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
-## v1.33.0-beta.2 - 2026-08-27
+## v1.34.0-beta.1 - 2026-08-29
+
+### 🚀 Features
+
+- **groups**: verwaltungs-user in der oberflaeche zuweisen (8e3f02c7)
+
+### 🐛 Bugfixes
+
+- **demo**: demo-manager der gruppe produktion zuweisen (7eac41ef)
+
+### 🔧 Sonstiges
+
+- **readme**: logo im kopfbereich, social-preview fuers repo (fc66f638)
+
+## v1.33.0 - 2026-08-27
 
 ### 🚀 Features